Changes for 64bit clean printf format strings.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@36810 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSSocketStream.m
@@ -596,7 +596,7 @@ static NSArray  *keys = nil;
 - (void) stream: (NSStream*)stream handleEvent: (NSStreamEvent)event
 {
   NSDebugMLLog(@"NSStream",
-    @"GSTLSHandler got %d on %p", event, stream);
+    @"GSTLSHandler got %"PRIdPTR" on %p", event, stream);
 
   if (handshake == YES)
     {
@@ -862,7 +862,7 @@ static NSString * const GSSOCKSAckConn = @"GSSOCKSAckConn";
 	  address = [[NSString alloc] initWithUTF8String:
 	    (char*)inet_ntoa(addr->sin_addr)];
 	  port = [[NSString alloc] initWithFormat: @"%d",
-	    (NSInteger)GSSwapBigI16ToHost(addr->sin_port)];
+	    (int)GSSwapBigI16ToHost(addr->sin_port)];
 
           /* Now reconfigure the streams so they will actually connect
            * to the socks proxy server.
